pyproject.toml
[tool.poetry]
name = "roboachxy"
version = "0.1.0"
description = "An discord bot for my friend groups"
authors = ["Achxy "]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 110
